# optimize get_smallest_vectors once more with numpy broadcasting and C