Despite the algorithmic improvements in recent commits, `get_smallest_vectors` is once again back at the top of the profiler results.  Unfortunately, I was not able to come up with any more algorithmic improvements, so I instead had to resort to careful use of numpy features to maximize data-crunching potential.

One last bit was difficult to implement using features available in numpy, so I wrote it in C instead.

---

On [these input files](https://github.com/ExpHP/phonopy-eigenvector-tblg-benchmark/tree/605f0d662fa8/input/big), using `--readfc`:
- **before:**
  - `total runtime`: 138.351s
  - `_get_smallest_vectors`: **73.126s** (53% of total)
- **after:**
  - `total runtime`: 71.632s
  - `_get_smallest_vectors`: **5.642s** (8% of total)

For reference, `eigh` (the "real work") takes 20.794s on these inputs.