@ExpHP 2017-08-15
Also, a note on the speed of sqrt:

Prior to optimizing the function, I instrumented it with IIFEs [as seen here](https://gist.github.com/ExpHP/0fd2e2d3b0d5e74b9d2de9fb8d834255) so I could get more detailed results from the profiler.

The breakdown was:

```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  1192464    1.997    0.000   24.164    0.000 cells.py:501(get_lattice_points)
  1192464   14.287    0.000   20.557    0.000 cells.py:528(filtered)
  1192464    1.244    0.000   13.260    0.000 cells.py:520(do_sums)
  1192464    6.106    0.000   13.040    0.000 cells.py:587(fill_it)
  1192464    3.888    0.000    7.590    0.000 cells.py:516(do_squares)
  1192464    6.267    0.000    6.267    0.000 cells.py:509(do_addition)
  1192464    2.252    0.000    2.252    0.000 cells.py:524(do_roots)  <------
        1    0.000    0.000    0.008    0.008 cells.py:553(prep)
```

Square roots really are surprisingly fast!  Allocations and conditional jumps (e.g. `filtered`) were much worse offenders.  (and of course, the lattice_points array that I should have made a global constant)