@atztogo 2017-08-16
Thanks @ExpHP. I could read your comments, but could not have time to reply because I wan in travel. I will review this PR tomorrow.